x86_64 port

+++ kernel/hal/x86_64/include/pmm.hpp
@@ -5,7 +5,7 @@
 #include <multiboot.h>
 namespace MTGosHAL {
 struct malloc_t {
-  uint32_t len;
+  size_t len;
   malloc_t *last;
   malloc_t *next;
 };
@@ -19,7 +19,7 @@ public:
   auto init(struct multiboot_info*) -> void;
   auto alloc(size_t length) -> void *;
   auto free(void* ptr) -> bool;
-  auto markUsed(const void * addr, uint32_t length) -> bool;
+  auto markUsed(const void * addr, uint64_t length) -> bool;
   auto operator >> (void * &addr) -> PMM &; //alloc
   auto operator << (const void * addr) -> PMM &; //free
   auto operator()(int pages) -> void*; //alloc_multipage
