+++ kernel/kernel/include/pmm.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 namespace MTGosHAL {
 struct malloc_t {
-  uint32_t len;
+  size_t len;
   malloc_t *last;
   malloc_t *next;
 };
@@ -12,7 +12,7 @@ private:
   malloc_t *head;
 public:
   PMM();
-  auto alloc(uint32_t length) -> void *;
+  auto alloc(size_t length) -> void *;
   auto free(void* ptr) -> bool;
   auto markUsed(const void * addr, uint32_t length) -> bool;
   auto operator >> (void * &addr) -> PMM &; //alloc
